requirements.txt
python-telegram-bot
astral
pandas
pexels-api